+++ crates/zng-app/src/crash_handler.rs
@@ -7,7 +7,7 @@
 use parking_lot::Mutex;
 use std::{
     fmt,
-    io::{BufRead, BufReader},
+    io::{BufRead, Write},
     path::{Path, PathBuf},
     time::SystemTime,
 };
@@ -15,10 +15,23 @@ use zng_layout::unit::TimeUnits as _;
 
 use zng_txt::{ToTxt as _, Txt};
 
+/// Environment variable that causes the crash handler to not start if set.
+///
+/// This is particularly useful to set in debugger launch configs. Crash handler spawns
+/// a different process for the app  so break points will not work.
+pub const NO_CRASH_HANDLER: &str = "ZNG_NO_CRASH_HANDLER";
+
 zng_env::on_process_start!(|process_start_args| {
+    if std::env::var(NO_CRASH_HANDLER).is_ok() {
+        return;
+    }
+
     let mut config = CrashConfig::new();
     for ext in CRASH_CONFIG {
         ext(&mut config);
+        if config.no_crash_handler {
+            return;
+        }
     }
 
     if std::env::var(APP_PROCESS) != Err(std::env::VarError::NotPresent) {
@@ -121,6 +134,7 @@ pub struct CrashConfig {
     app_process: ConfigProcess,
     dialog_process: ConfigProcess,
     dump_dir: Option<PathBuf>,
+    no_crash_handler: bool,
 }
 impl CrashConfig {
     fn new() -> Self {
@@ -130,6 +144,7 @@ impl CrashConfig {
             app_process: vec![],
             dialog_process: vec![],
             dump_dir: Some(zng_env::cache("zng_minidump")),
+            no_crash_handler: false,
         }
     }
 
@@ -179,6 +194,13 @@ impl CrashConfig {
     /// Do not collect a minidump.
     pub fn no_minidump(&mut self) {
         self.dump_dir = None;
+    }
+
+    /// Does not run with crash handler.
+    ///
+    /// This is equivalent of running with `NO_ZNG_CRASH_HANDLER` env var.
+    pub fn no_crash_handler(&mut self) {
+        self.no_crash_handler = true;
     }
 }
 
@@ -989,33 +1011,34 @@ fn run_process(command: &mut std::process::Command) -> std::io::Result<(std::pro
 
     Ok((status, [stdout, stderr]))
 }
-fn capture_and_print(stream: impl std::io::Read + Send + 'static, is_err: bool) -> std::thread::JoinHandle<String> {
+fn capture_and_print(mut stream: impl std::io::Read + Send + 'static, is_err: bool) -> std::thread::JoinHandle<String> {
     std::thread::spawn(move || {
-        let mut capture = String::new();
-
-        let mut reader = BufReader::new(stream);
-
+        let mut capture = vec![];
+        let mut buffer = [0u8; 32];
         loop {
-            let mut line = String::new();
-            match reader.read_line(&mut line) {
+            match stream.read(&mut buffer) {
                 Ok(n) => {
-                    if n > 0 {
-                        if is_err {
-                            eprint!("{line}");
-                        } else {
-                            print!("{line}");
-                        }
-                        capture.push_str(&line);
-                        line.clear();
-                    } else {
+                    if n == 0 {
                         break;
+                    }
+
+                    let new = &buffer[..n];
+                    capture.write_all(new).unwrap();
+                    let r = if is_err {
+                        let mut s = std::io::stderr();
+                        s.write_all(new).and_then(|_| s.flush())
+                    } else {
+                        let mut s = std::io::stdout();
+                        s.write_all(new).and_then(|_| s.flush())
+                    };
+                    if let Err(e) = r {
+                        panic!("{} write error, {}", if is_err { "stderr" } else { "stdout" }, e)
                     }
                 }
                 Err(e) => panic!("{} read error, {}", if is_err { "stderr" } else { "stdout" }, e),
             }
         }
-
-        capture
+        String::from_utf8_lossy(&capture).into_owned()
     })
 }